Move report validation
