+++ katprep/reports.py
@@ -1,4 +1,7 @@
 import json
+import os
+import os.path
+from argparse import ArgumentTypeError
 
 from . import get_json
 from .host import Host
@@ -37,3 +40,35 @@ def write_report(filename, report):
 
     with open(filename, "w") as target:
         target.write(json.dumps(report, default=convert_hosts))
+
+
+def is_valid_report(filename):
+    """
+    Checks whether a JSON file contains a valid snapshot report.
+
+    :param filename: the JSON filename
+    :type filename: str
+    """
+    if not os.path.exists(filename):
+        raise ArgumentTypeError("File '{filename!r}' is non-existent")
+
+    if not os.access(filename, os.R_OK):
+        raise ArgumentTypeError("File '{filename!r}' is not readable")
+
+    try:
+        # check whether valid json
+        json_obj = load_report(filename)
+
+        if not isinstance(json_obj, dict):
+            raise TypeError("The report contains the wrong root type"
+                            " -  expected an object (dict).")
+
+        if not json_obj:
+            raise ValueError("The report is empty")
+    except (TypeError, ValueError) as err:
+        raise ArgumentTypeError(f"File {filename} is not a valid JSON"
+                                f" snapshot report: {err}")
+    except Exception as err:
+        raise ArgumentTypeError(f"File {filename} failed to load: {err}")
+
+    return filename
